Move model classes of ODL switchManager from Client to Model bundle.
