+++ extensions/odl/capabilities/src/main/java/org/mqnaas/extensions/odl/capabilities/impl/ODLRootResourceProvider.java
@@ -50,11 +50,11 @@ import org.mqnaas.core.api.exceptions.ResourceNotFoundException;
 import org.mqnaas.core.impl.AttributeStore;
 import org.mqnaas.core.impl.RootResource;
 import org.mqnaas.extensions.odl.client.switchnorthbound.ISwitchNorthboundAPI;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.Node.NodeType;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeConnectorProperties;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeConnectors;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeProperties;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.Nodes;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectorProperties;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectors;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeProperties;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.Nodes;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node.NodeType;
 import org.mqnaas.network.api.topology.port.IPortManagement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
