+++ extensions/odl/capabilities/src/test/java/org/mqnaas/extensions/odl/capabilities/impl/ODLRootResourceProviderTest.java
@@ -56,14 +56,14 @@ import org.mqnaas.core.api.exceptions.ResourceNotFoundException;
 import org.mqnaas.core.impl.AttributeStore;
 import org.mqnaas.core.impl.RootResource;
 import org.mqnaas.extensions.odl.client.switchnorthbound.ISwitchNorthboundAPI;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.Node;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.Node.NodeType;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeConnector;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeConnectorProperties;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeConnectors;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.NodeProperties;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.Nodes;
-import org.mqnaas.extensions.odl.client.switchnorthbound.api.PropertyValue;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node.NodeType;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnector;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectorProperties;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectors;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeProperties;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.Nodes;
+import org.mqnaas.extensions.odl.hellium.switchmanager.model.PropertyValue;
 import org.mqnaas.general.test.helpers.reflection.ReflectionTestHelper;
 import org.mqnaas.network.api.topology.port.IPortManagement;
 import org.mqnaas.network.impl.topology.port.PortManagement;
