+++ extensions/odl/model/src/main/java/org/mqnaas/extensions/odl/hellium/switchmanager/model/PropertyValue.java
@@ -1,4 +1,4 @@
-package org.mqnaas.extensions.odl.client.switchnorthbound.api;
+package org.mqnaas.extensions.odl.hellium.switchmanager.model;
 
 /*
  * #%L
@@ -21,44 +21,56 @@ package org.mqnaas.extensions.odl.client.switchnorthbound.api;
  * #L%
  */
 
-import java.util.List;
-
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
 /**
- * {@link NodeConnectorProperties} wrapper.
- * 
- * Based on OpenDaylight Helium release (<a href=
- * "https://github.com/opendaylight/controller/blob/stable/helium/opendaylight/northbound/switchmanager/src/main/java/org/opendaylight/controller/switchmanager/northbound/NodeConnectors.java"
- * >reference</a>).
+ * {@link NodeProperties} property value with mandatory value and optional name attributes.
  * 
  * @author Julio Carlos Barrera
  *
  */
-@XmlRootElement(name = "list")
+@XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
-public class NodeConnectors {
+public class PropertyValue {
 
-	private List<NodeConnectorProperties>	nodeConnectorProperties;
+	private String	value;
 
-	public NodeConnectors() {
+	@XmlElement(required = false)
+	private String	name;
+
+	public PropertyValue() {
 	}
 
-	public List<NodeConnectorProperties> getNodeConnectorProperties() {
-		return nodeConnectorProperties;
+	public PropertyValue(String value, String name) {
+		this.value = value;
+		this.name = name;
 	}
 
-	public void setNodeConnectorProperties(List<NodeConnectorProperties> nodeConnectorProperties) {
-		this.nodeConnectorProperties = nodeConnectorProperties;
+	public String getValue() {
+		return value;
+	}
+
+	public void setValue(String value) {
+		this.value = value;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
 	}
 
 	@Override
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + ((nodeConnectorProperties == null) ? 0 : nodeConnectorProperties.hashCode());
+		result = prime * result + ((name == null) ? 0 : name.hashCode());
+		result = prime * result + ((value == null) ? 0 : value.hashCode());
 		return result;
 	}
 
@@ -70,18 +82,23 @@ public class NodeConnectors {
 			return false;
 		if (getClass() != obj.getClass())
 			return false;
-		NodeConnectors other = (NodeConnectors) obj;
-		if (nodeConnectorProperties == null) {
-			if (other.nodeConnectorProperties != null)
+		PropertyValue other = (PropertyValue) obj;
+		if (name == null) {
+			if (other.name != null)
 				return false;
-		} else if (!nodeConnectorProperties.equals(other.nodeConnectorProperties))
+		} else if (!name.equals(other.name))
+			return false;
+		if (value == null) {
+			if (other.value != null)
+				return false;
+		} else if (!value.equals(other.value))
 			return false;
 		return true;
 	}
 
 	@Override
 	public String toString() {
-		return "NodeConnectors [nodeConnectorProperties=" + nodeConnectorProperties + "]";
+		return "PropertyValue [value=" + value + ", name=" + name + "]";
 	}
 
 }
