+++ extensions/odl/model/src/test/resources/serialization/test-node-connectors.xml
